New logic for requesting storage permissions (Android) on start up if needed. (requires optimization in the messages generation and maybe extract of some functions/widgets to its own file).
The app will load a blank screen with the name in the middle the first time, until the permission is granted, or if the permission is denied by the user.
Includes several dialogs to explain the user why the permission is required and what happen if it is not granted.