Preview action implemented. When previewing a file, a timer is set for 5 seconds before deleting the temporary file.
This is because, unlike the sharing action, the file needs to be present when the default app
tries to open it, and when the user needs to pick the app, if no default is set in the phone,
it can take some seconds for this to happen.

Also: Code cleanup and refactoring.