A BaseItem parameter was added to the RouteBloc and NavigationBloc (event, states), to pass the data of the item in each navigation.
If the item tapped is a file, then instead of firing a NavigateTo event, a MaterialPageRoute is pushed for the FilePage. This way there is not issue with the bloc instance stopping firing events if the push is performed using the bloc (after returning from filepage.dart, the events wouldn't fire again).

If the item is a folder, a NavigateTo event is fired.