Repository notification subscription
The ouisync repository was open when required, then closed right away, in the DirectoryRepository.
In order to subscribe to notifications in the ouisync repository, now an instance is open at the app startup, then injected into the DirectoryRepository.

This means the ouisync session is not needed everywhere to pass into the DirectoryRepository in order to open an ouisync repository, so it was removed from all the places in which it was passed as a parameter to the bloc.

Also: imports fixed/ordered, small code cleanups.