Bug fix: navigation to created folder was not refreshing the contents
In the directory bloc state, the path of the directory was added. When a directory load state (DirectoryLoadSuccess) generated from the navigation coincides with another coming from the syncing event, the path is used to check agains the current folder and decide if the contents needs to be updated.

Also: the SyncingInProgress state was excluded from the directory bloc builder, to avoid unecessary calls to replace the layout.