Modal bottom sheets for folders and files details updated
The way to move a folder/file was updated to use a permanent bottom sheet instead of a dropdown with the folders in the repository.
A callback for updating the controller for the permanent bottom sheet, and a callback for moving the folder, both located in the main_page.dart, are passed in via the constructor; also, a global key for the scaffold is passed in the constructor for showing the persistent bottom sheet.