Bug fix: An error state was shown when the current repository was syncing, while locked (blind replica)
With the repository locked, the syncing process tried to get the contents of the current folder, which returned
an error from the library: permission denied. Then an error state was returned from the DirectoryLoadingFailure
event.

To fix it, getting the contents of a folder during the syncing process, is only performed if the repository
is unlocked (access mode != blind)

Fixes #60