Cancel file saving
In the file_description.dart widget, a 'CANCEL' button was added at the end of the progress bar, to allow the user to do this.
In the DirectoryBloc, the function _onSaveFile was modified, so while writing to the file via an await for loop, if the user
taps the 'CANCEL' button,  the bloc receives a new event of type CancelSaveFile and changes the private flag _cancelFileWriting,
which cancel the loop that is using a stream created via a takeWhile function.

Also:  main_page.dart updated to show a toaster notifying the cancelation and other small refactoring.