Fixed: function saveMedia was popping the top widget in the navigator, even when no modal was present
For adding files via a share intent, a modal is used to confirm the destination folder, and it needs to be popped when done.
When files are added using drag and drop, this modal is not used, hence popping was removing the main page.

A flag was added and made true from inside the modal, using a callback function, to determine when to pop.