Fixed: Cancel validation for downloading file
The validatiion for canceling a downloading file was using only the
file identifier, which caused that if a file with
the same name was being downloaded at the same time in two or more
repositories, then all of them would be canceled.

Now both the repository and the file identifier are used for the validation.