Fixed: Cancel validation for saving a file
The validatiion for canceling saving a file was using only the
file identifier, which caused that if a file with
the same name was being saved at the same time in two or more
repositories, then all of them would be canceled.

Now both the repository and the file identifier are used for the validation.