Enable autoVerify for App Links
To open share tokens in the OuiSync App without showing the
disambiguator. Note that it'll work only for our app when signed by the
release key. I.e. it won't work in debug because we don't have the debug
signing key stored on git.

To make this work we also had to do this change in ouisync-web:

https://github.com/equalitie/ouisync-web/commit/082937a6a1dca6bd519560d8defa80516c8d8175