Fixed: repo list overflowed when needing scroll
When the list of repos contained too many entries, the list would grow,
causing the bottom sheet to grow as well.

When there was not more space to grow, the list would still try to grow,
overflowing vertically.

For this, the list was placed inside a constrained box, which only
allows it to grow up to 40% of the screen height.