Fixed: No repo state when loading app
When picking the mainn state when loading the app, the only option when
there was not default repo selected was the state for no repositories.

Now an empty widget is used everytime the current repo is null, but
there is at least one repo, instead of just use the no repos state;
right after this empty widget a loading state is used, which is the
reason an extra state was not created.