Fixed: removing biometrics locked the user out of the repo
When removing the biometrics (version2 => no_password), the password
saved in the secured storage was still in the entry with the key based
in the version2 authentication mode, this is: databaseId-v2

Because of this, when changing the repo from version2 to
no_local_password, the entry could not be retrieved.

To fix this, everytime a repo is updated from or to version2, the
entry in the secured storage is replaced with a new one.