[iOS] Fix: ApplicationDocumentsDirectory on debug
In iOS, if the app is build as debug, the path to the app directories
changes. Mlore especifically the path contains a section that
corresponds to the sandbox for the debug sesion, which is different in
every case (this only happens on Debug).

Because of this, we can't depend on saving the path the first time we
adquire it, and use it every time we run the app, and instead we check
 if it was compiled as debug, using the kDebugMode constant, and each
 time adquiring the path we need.