Fix: showMessage fired by RepoCubit more than once
The listener in the bloc consumer on main_page.dart was only checking for the message field on the state being no empty.
Because of this, next time the cubit emited a new state, the listener would display the mesage again, since the message field was never cleaned.

The message field was removed from the state, as well as the showMessage; instead we used showSnackBar as needed.

Related: issue #393