Security screen fix: password
Fix a bug where after "updating" a password in the repository screen and
then leaving there was a popup about unsaved changes. This was because
the password was still in the password field and the old logic evaluated
it as a pending change. Fixed by introducing another field
`updatedLocalPassword` to signify the latest set password and checking
if the current password in the password field is equal to that.