Remove incorrect warning text from the security screen
This happened when the user accessed the security screen while the
repository was locked. The message was "The security options are not
available for blind repositories", but given that the user did in fact
enter the security screen means that they had a password to the repo and
thus it was a locked repo as opposed to a blind one.