Drop iOS, consolidate Mac versions
This drops iOS support because we have a separate flutter target for that which
is likely going to use a different interface (the one we're using here is only
available on iOS 16 and above).

Also sets the minimum target version to 13.0 because this is the lowest version
I was able to compile it on (well, 13.6.7 if we're splitting hairs but it
probably doesn't matter because...). If desired, we can bring this as low as
11.0 since the only changes required are developer-experience related (we're
using some URL pathing features that were not backported and we'd have to
sleep in nanoseconds).