Adopt linux behavior for macOS tray icon
Like linux, macOS renders the title inline with the tray icon. This is not very
useful for a static title (it might be useful to, for example, show live
bandwidth stats instead) so it's currently disabled.

Because right clicking in the tray area is uncommon on mac, we also adopt the
menu functionality from linux with separate show / hide and exit.

Additionally, flutter's docs state that `windowManager.destroy()` should be
called to really close the window and I can confirm that it is required on
macOS, otherwise the window stays open.