Fix log path on macOS
The logs were not visible on mac due to a divergence in the log file paths.
This attempts to fix it by consolidating the locations in the flutter app. It's
a stopgap solution, I'd prefer if we had a centralized storage location for all
things related to paths that is also customizable from the host side (a local
path_provider module, if you will).