Add support for async trace demangling
Flutter doesn't natively know how to parse a Chain of async stack
traces and recommends setting this property to do so.

Personally, I've had mixed results with this: on one hand, it allows
Flutter to trim a large number of `rebuild` calls, but on the other it
seems to delete potentially useful context, which is why I've made this
optional based on the existence of an environment variable.