Throttle session restarts on error
This addresses two related issues. The first was that in the event that
the session is invalidated, the app continues using the old session until
a new one is established. This is corrected by nullifying the session
when this happens. The second was that retries were never attempted if
the session is unable to initialize in the first place. To ensure that we
don't end up spinning on the CPU in the event that a session cannot be
established in the first place, a minimum delay of 1 second is now in
effect between any two session initializations.